agent/crewai-flow-multimodal/pyproject.toml
[project]
name = "crewai-flow-multimodal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.69.0",
    "assemblyai>=0.44.3",
    "crewai>=0.201.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "llama-index-embeddings-ollama>=0.8.3",
    "ollama>=0.6.0",
    "pymilvus>=2.6.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
]
